requirements.txt
groq
python-dotenv
tabulate
pytest
